pyproject.toml
[project]
name = "clip-edit-tool"
version = "0.1.0"
description = "YouTube video repurposing application with AI-powered clipping"
requires-python = ">=3.8"
dependencies = [
    "pytubefix>=9.2.0",
]

[tool.uv]
dev-dependencies = []